requirements.txt
Flask==2.3.3
yfinance==0.2.28
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
supabase==1.1.1
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0
httpx==0.24.1